Create a non-entity test module
