+++ tests/modules/nonentity_pathauto_test/contact_pathauto_test.module
@@ -1,0 +1,33 @@
+<?php
+
+/**
+ * @file
+ * Contains contact_pathauto_test.module.
+ */
+
+use Drupal\Core\Routing\RouteMatchInterface;
+
+/**
+ * Implements hook_help().
+ */
+function contact_pathauto_test_help($route_name, RouteMatchInterface $route_match) {
+  switch ($route_name) {
+    // Main module help for the contact_pathauto_test module.
+    case 'help.page.contact_pathauto_test':
+      $output = '';
+      $output .= '<h3>' . t('About') . '</h3>';
+      $output .= '<p>' . t('A test module for non-entity pathauto generation') . '</p>';
+      return $output;
+
+    default:
+  }
+}
+
+/**
+ * Implements hook_theme().
+ */
+function contact_pathauto_test_theme() {
+  $theme = [];
+
+  return $theme;
+}
